+++ config/elastic.exs
@@ -48,5 +48,34 @@ config :td_dq, TdDq.Search.Cluster,
 
       # Support create or replace
       bulk_action: "index"
+    },
+    implementations: %{
+       # This map describes the mappings and settings for your index. It will
+      # be posted as-is to Elasticsearch when you create your index, and
+      # therefore allows all the settings you could post directly.
+      settings: %{},
+
+      # This store module must implement a store behaviour. It will be used to
+      # fetch data for each source in each indexes' `sources` list, below:
+      store: TdDq.Search.Store,
+
+      # This is the list of data sources that should be used to populate this
+      # index. The `:store` module above will be passed each one of these
+      # sources for fetching.
+      #
+      # Each piece of data that is returned by the store must implement the
+      # Elasticsearch.Document protocol.
+      sources: [TdDq.Rules.Implementations.Implementation],
+
+      # Controls the data ingestion rate by raising or lowering the number
+      # of items to send in each bulk request.
+      bulk_page_size: 1000,
+
+      # Likewise, wait a given period between posting pages to give
+      # Elasticsearch time to catch up.
+      bulk_wait_interval: 0,
+
+      # Support create or replace
+      bulk_action: "index"
     }
   }