+++ lib/td_dq/rules/rule_search.ex
@@ -2,74 +2,86 @@ defmodule TdDq.Rules.Search do
   require Logger
 
   @moduledoc """
-    Helper module to construct rule search queries.
+    Helper module to construct search queries.
   """
   alias TdDq.Accounts.User
   alias TdDq.Permissions
   alias TdDq.Search
-  alias TdDq.Search.Aggregations
   alias TdDq.Search.Query
 
-  def get_filter_values(%User{is_admin: true}, params) do
-    filter_clause = Query.create_filters(params)
+  def get_filter_values(user, params, index \\ :rules)
+
+  def get_filter_values(%User{is_admin: true}, params, index) do
+    filter_clause = Query.create_filters(params, index)
     query = Query.create_query(%{}, filter_clause)
-    search = %{query: query, aggs: Aggregations.aggregation_terms()}
-    Search.get_filters(search)
+    search = %{query: query, aggs: Query.get_aggregation_terms(index)}
+    Search.get_filters(search, index)
   end
 
-  def get_filter_values(%User{} = user, params) do
-    user_defined_filters = Query.create_filters(params)
+  def get_filter_values(%User{} = user, params, index) do
+    user_defined_filters = Query.create_filters(params, index)
 
     permissions =
       user
       |> Permissions.get_domain_permissions()
-      |> rule_permissions(user_defined_filters)
+      |> get_permissions(user_defined_filters, index)
 
-    get_filter_values(permissions, params)
+    get_filters(permissions, params, index)
   end
 
-  def get_filter_values([], _), do: %{}
+  def search(params, user, page \\ 0, size \\ 50, index \\ :rules)
 
-  def get_filter_values(permissions, params) do
-    user_defined_filters = Query.create_filters(params)
-    user_defined_filters = user_defined_filters |> delete_execution_filter
-    filter = Query.create_filter_clause(permissions, user_defined_filters)
-    query = Query.create_query(params, filter)
-    search = %{query: query, aggs: Aggregations.aggregation_terms()}
-    Search.get_filters(search)
-  end
-
-  def search(params, user, page \\ 0, size \\ 50)
-
-  def search(params, %User{is_admin: true}, page, size) do
-    filter_clause = Query.create_filters(params)
+  def search(params, %User{is_admin: true}, page, size, index) do
+    filter_clause = Query.create_filters(params, index)
     filter_clause = filter_clause |> delete_execution_filter
     query = Query.create_query(params, filter_clause)
 
-    sort = Map.get(params, "sort", ["name.raw"])
+    sort = Map.get(params, "sort", default_sort(index))
 
     %{
       from: page * size,
       size: size,
       query: query,
       sort: sort,
-      aggs: Aggregations.aggregation_terms()
+      aggs: Query.get_aggregation_terms(index)
     }
-    |> do_search
+    |> do_search(index)
   end
 
-  def search(params, %User{} = user, page, size) do
-    user_defined_filters = Query.create_filters(params)
+  def search(params, %User{} = user, page, size, index) do
+    user_defined_filters = Query.create_filters(params, index)
 
     permissions =
       user
       |> Permissions.get_domain_permissions()
-      |> rule_permissions(user_defined_filters)
+      |> get_permissions(user_defined_filters, index)
 
-    filter_rules(params, permissions, page, size)
+    filter_rules(params, permissions, page, size, index)
   end
 
-  defp rule_permissions(domain_permissions, user_defined_filters) do
+  defp get_filters([], _, _index), do: %{}
+
+  defp get_filters(permissions, params, index) do
+    user_defined_filters = Query.create_filters(params, index)
+    user_defined_filters = user_defined_filters |> delete_execution_filter
+    filter = Query.create_filter_clause(permissions, user_defined_filters)
+    query = Query.create_query(params, filter)
+    search = %{query: query, aggs: Query.get_aggregation_terms(index)}
+    Search.get_filters(search, index)
+  end
+
+  defp default_sort(:rules), do: ["name.raw"]
+
+  defp default_sort(:implementations), do: ["implementation_key.raw"]
+
+  defp get_permissions(domain_permissions, user_defined_filters, index) do
+    case index do
+      :rules -> get_permissions(domain_permissions, user_defined_filters)
+      :implementations -> get_permissions(domain_permissions)
+    end
+  end
+
+  defp get_permissions(domain_permissions, user_defined_filters) do
     domain_permissions
     |> Enum.filter(fn permissions_obj ->
       case do_rules_execution(user_defined_filters) do
@@ -79,6 +91,12 @@ defmodule TdDq.Rules.Search do
         false ->
           Enum.any?(permissions_obj.permissions, &check_view_or_manage_permission(&1))
       end
+    end)
+  end
+
+  defp get_permissions(domain_permissions) do
+    Enum.filter(domain_permissions, fn %{permissions: permissions} ->
+      Enum.any?(permissions, &check_view_or_manage_permission(&1))
     end)
   end
 
@@ -92,11 +110,11 @@ defmodule TdDq.Rules.Search do
       permission_names == :manage_confidential_business_concepts
   end
 
-  defp filter_rules(_params, [], _page, _size),
+  defp filter_rules(_params, [], _page, _size, _index),
     do: %{results: [], aggregations: %{}, total: 0}
 
-  defp filter_rules(params, [_h | _t] = permissions, page, size) do
-    user_defined_filters = Query.create_filters(params)
+  defp filter_rules(params, [_h | _t] = permissions, page, size, index) do
+    user_defined_filters = Query.create_filters(params, index)
     user_defined_filters = user_defined_filters |> delete_execution_filter
     filter = Query.create_filter_clause(permissions, user_defined_filters)
 
@@ -109,31 +127,21 @@ defmodule TdDq.Rules.Search do
       size: size,
       query: query,
       sort: sort,
-      aggs: Aggregations.aggregation_terms()
+      aggs: Query.get_aggregation_terms(index)
     }
-    |> do_search
+    |> do_search(index)
   end
 
-  defp do_search(search) do
-    %{results: results, aggregations: aggregations, total: total} = Search.search(search)
+  defp do_search(search, index) do
+    %{results: results, aggregations: aggregations, total: total} = Search.search(search, index)
 
     results =
       results
       |> Enum.map(&Map.get(&1, "_source"))
-      |> Enum.map(&atomize_keys/1)
+      |> Enum.map(&Map.Helpers.atomize_keys/1)
 
     %{results: results, aggregations: aggregations, total: total}
   end
-
-  defp atomize_keys(%{} = map) do
-    map
-    |> Enum.into(%{}, fn {k, v} -> {atomize_key(k), atomize_keys(v)} end)
-  end
-
-  defp atomize_keys(value), do: value
-
-  defp atomize_key(key) when is_binary(key), do: String.to_atom(key)
-  defp atomize_key(key), do: key
 
   defp do_rules_execution(user_defined_filters) do
     user_defined_filters