+++ lib/td_dq/scheduler.ex
@@ -1,0 +1,5 @@
+defmodule TdDq.Scheduler do
+  @moduledoc "A Quantum scheduler for launching periodic tasks"
+
+  use Quantum, otp_app: :td_dq
+end