+++ lib/td_dq_web/views/filter_view.ex
@@ -1,4 +1,4 @@
-defmodule TdDqWeb.RuleFilterView do
+defmodule TdDqWeb.FilterView do
   use TdDqWeb, :view
 
   def render("show.json", %{filters: filters}) do